Updated to the newest version of GStreamer with all depending libraries. The internal audio player is working now

+++ src/control/VolumeManager.java
@@ -73,7 +73,7 @@ public class VolumeManager
 	 * 
 	 * @return Volume in percent
 	 */
-	public int getVolume()
+	public double getVolume()
 	{
 		if(list.capacity() > 0)
 		{
